moved SanitizeName to avoid circular dependency
