+++ x509util/utils.go
@@ -11,6 +11,7 @@ import (
 	"encoding/asn1"
 	"encoding/base64"
 	"encoding/hex"
+	"golang.org/x/net/idna"
 	"math/big"
 	"net"
 	"net/url"
@@ -33,6 +34,27 @@ const (
 )
 
 var emptyASN1Subject = []byte{0x30, 0}
+
+func SanitizeName(domain string) (string, error) {
+	if domain == "" {
+		return "", errors.New("empty server name")
+	}
+
+	// Note that this conversion is necessary because some server names in the handshakes
+	// started by some clients (such as cURL) are not converted to Punycode, which will
+	// prevent us from obtaining certificates for them. In addition, we should also treat
+	// example.com and EXAMPLE.COM as equivalent and return the same certificate for them.
+	// Fortunately, this conversion also helped us deal with this kind of mixedcase problems.
+	//
+	// Due to the "σςΣ" problem (see https://unicode.org/faq/idn.html#22), we can't use
+	// idna.Punycode.ToASCII (or just idna.ToASCII) here.
+	name, err := idna.Lookup.ToASCII(domain)
+	if err != nil {
+		return "", errors.New("server name contains invalid character")
+	}
+
+	return name, nil
+}
 
 // SplitSANs splits a slice of Subject Alternative Names into slices of
 // IP Addresses and DNS Names. If an element is not an IP address, then it
